requirements.txt
Flask>=3.1,<4.0
aiortc>=1.9,<2.0
pytest>=8.0,<9.0
playwright>=1.52,<2.0

google-genai>=0.5.0
requests>=2.31.0
python-dotenv>=1.0,<2.0
Pillow>=10.0,<12.0
